Minor updates to the French localisation
Using the more standard word "Ouvrir" to mean "Open" an app.

Note: the "Launch at login" may be a bit long in this version, but this is following standard phrasing that macOS uses in System Preferences. Also with this version there is a repetition of words stemmed from "ouvrir/ouverture". 
Another option is: "Ouvrir avec la session" or "Ouvrir au démarrage"